move some classes to jx_base
