+++ jx_base/__init__.py
@@ -2,13 +2,22 @@
 #
 #
 # This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with self file,
-# You can obtain one at http:# mozilla.org/MPL/2.0/.
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at http://mozilla.org/MPL/2.0/.
 #
 # Author: Kyle Lahnakoski (kyle@lahnakoski.com)
 #
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import unicode_literals
 
 from uuid import uuid4
+
+
+OBJECT = "object"
+NESTED = "nested"
+STRUCT = [OBJECT, NESTED]
+
 
 def generateGuid():
     """Gets a random GUID.
