+++ jx_python/cubes/aggs.py
@@ -17,7 +17,7 @@ from mo_collections.matrix import Matrix
 from mo_logs import Log
 from mo_dots import listwrap
 from jx_python import windows
-from jx_python.domains import SimpleSetDomain, DefaultDomain
+from jx_base.domains import SimpleSetDomain, DefaultDomain
 from jx_base.expressions import jx_expression_to_function
