+++ jx_python/group_by.py
@@ -18,8 +18,8 @@ import sys
 from mo_dots import listwrap, Null, Data
 from mo_logs import Log
 
+from jx_base.container import Container
 from jx_base.expressions import jx_expression_to_function, jx_expression, Expression
-from jx_python.containers import Container
 from mo_collections.multiset import Multiset
 from mo_dots.lists import FlatList
 from mo_logs.exceptions import Except
