+++ jx_python/lists/aggs.py
@@ -13,15 +13,16 @@ from __future__ import unicode_literals
 
 import itertools
 
+from jx_python import windows
+from mo_collections.matrix import Matrix
+from mo_dots import listwrap, wrap, coalesce
 from mo_logs import Log
 from mo_math import UNION
 from mo_times.dates import Date
-from mo_dots import listwrap, wrap, coalesce
-from mo_collections.matrix import Matrix
-from jx_python import windows
-from jx_python.domains import SimpleSetDomain, DefaultDomain
-from jx_python.expression_compiler import compile_expression
+
+from jx_base.domains import SimpleSetDomain, DefaultDomain
 from jx_base.expressions import jx_expression_to_function
+from jx_python.expression_compiler import compile_expression
 
 _ = Date
