+++ jx_python/namespace/__init__.py
@@ -14,7 +14,7 @@ from __future__ import absolute_import
 from collections import Mapping
 
 from mo_dots import set_default, Data
-from jx_python.query import QueryOp
+from jx_base.query import QueryOp
 
 
 class Namespace(object):
