+++ jx_sqlite/base_table.py
@@ -19,7 +19,7 @@ from jx_base import generateGuid
 from jx_sqlite import UID, GUID
 from jx_sqlite.snowflake import Snowflake
 from jx_python import jx
-from jx_python.containers import Container
+from jx_base.container import Container
 from pyLibrary.sql.sqlite import Sqlite
 
 _config=None
@@ -43,7 +43,7 @@ class BaseTable(Container):
 
         if not _config:
             # REGISTER sqlite AS THE DEFAULT CONTAINER TYPE
-            from jx_python.containers import config as _config
+            from jx_base.container import config as _config
             if not _config.default:
                 _config.default = {
                     "type": "sqlite",
