+++ jx_sqlite/insert_table.py
@@ -21,7 +21,7 @@ from mo_logs import Log
 
 from jx_sqlite import typed_column, quote_table, get_type, ORDER, UID, GUID, PARENT, get_if_type
 from jx_sqlite.base_table import BaseTable, generateGuid
-from jx_python.containers import STRUCT
+from jx_base import STRUCT
 from jx_base.expressions import jx_expression
 from jx_python.meta import Column
 from pyLibrary.sql.sqlite import quote_value, quote_column
@@ -236,7 +236,7 @@ class InsertTable(BaseTable):
             :param row: we will be filling this
             :return:
             """
-            table=concat_field(self.sf.fact, nested_path[0])            
+            table=concat_field(self.sf.fact, nested_path[0])
             insertion = doc_collection[nested_path[0]]
             if not row:
                 row = {GUID: guid, UID: uid, PARENT: parent_id, ORDER: order}
@@ -286,7 +286,7 @@ class InsertTable(BaseTable):
                     column = c.es_column
                     from_doc.active_columns.remove(c)
                     abs_schema.remove(cname, c)
-                    required_changes.append({"nest": (c, nested_path[0])})                    
+                    required_changes.append({"nest": (c, nested_path[0])})
                     deep_c = Column(
                         names={".": cname},
                         type=value_type,
@@ -302,7 +302,7 @@ class InsertTable(BaseTable):
                         if column in r1:
                             row1 = {UID: self.next_uid(), PARENT: r1["__id__"], ORDER: 0, column: r1[column]}
                             insertion.rows.append(row1)
-                            
+
                 elif len(c.nested_path) > len(nested_path):
                     insertion = doc_collection[c.nested_path[0]]
                     row = {UID: self.next_uid(), PARENT: uid, ORDER: order}
