+++ jx_sqlite/query_table.py
@@ -14,18 +14,19 @@ from __future__ import division
 from __future__ import unicode_literals
 
 import mo_json
-from jx_sqlite import quote_table, sql_aggs, unique_name, untyped_column, typed_column, json_types
+from jx_python import jx
+from jx_sqlite import quote_table, sql_aggs, unique_name, untyped_column
 from mo_collections.matrix import Matrix, index_to_coordinate
-from mo_dots import listwrap, coalesce, Data, wrap, startswith_field, unliteral_field, unwrap, split_field, join_field, unwraplist, concat_field, relative_field
+from mo_dots import listwrap, coalesce, Data, wrap, startswith_field, unwrap, unwraplist, concat_field, relative_field
 from mo_logs import Log
 
-from jx_sqlite.aggs_table import AggsTable
-from jx_python import jx
-from jx_python.containers import STRUCT
-from jx_python.domains import SimpleSetDomain
+from jx_base.domains import SimpleSetDomain
 from jx_base.expressions import jx_expression, Variable, TupleOp
-from jx_python.query import QueryOp
+from jx_base import STRUCT
 from jx_python.meta import Column
+from jx_base.query import QueryOp
+from jx_sqlite.aggs_table import AggsTable
+
 
 class QueryTable(AggsTable):
     def get_column_name(self, column):
@@ -147,7 +148,7 @@ class QueryTable(AggsTable):
 
                 return Data(
                     data=unwrap(data),
-                    select=select,                    
+                    select=select,
                     meta={"format": "cube"}
                 )
 
@@ -461,6 +462,6 @@ class QueryTable(AggsTable):
         return output
 
 
-from jx_python.containers import type2container
+from jx_base.container import type2container
 
 type2container["sqlite"] = QueryTable
