+++ jx_sqlite/setop_table.py
@@ -19,7 +19,7 @@ from mo_math import UNION, MAX
 from jx_base.queries import get_property_name
 from jx_sqlite import quote_table, quoted_UID, quoted_GUID, get_column, _make_column_name, ORDER, COLUMN, set_column, quoted_PARENT, ColumnMapping
 from jx_sqlite.insert_table import InsertTable
-from jx_python.containers import STRUCT
+from jx_base import STRUCT
 from jx_sqlite.expressions import sql_type_to_json_type, LeavesOp
 from jx_python.meta import Column
 from pyLibrary.sql.sqlite import quote_value
